Rename types -> _types, add __init__.py and fix circular imports
