+++ backend/modules/experiment.py
@@ -1,17 +1,17 @@
 """TODO document"""
 
-from types.experiment_state import ExperimentState
-from modules.experimenter import Experimenter
-from modules.participant import Participant
-from modules.session import Session
+from _types.experiment_state import ExperimentState
+import modules.experimenter as _experimenter
+import modules.participant as _participant
+import modules.session as _session
 
 
 class Experiment():
     """TODO document"""
     state: ExperimentState
-    session: Session
-    experimenters: list[Experimenter]
-    participants: list[Participant]
+    session: _session.Session
+    experimenters: list[_experimenter.Experimenter]
+    participants: list[_participant.Participant]
 
     def start(self):
         """TODO document"""
